Added CONFIG_TTY , CONFIG_VT
CONFIG_TTY
TTY - in order to have a   "Input from somewhere/Output from somewhere [e.g. keyboard and dispaly] you need to have a TTY". A fantastic explanation (especially the history!) is in https://www.howtogeek.com/428174/what-is-a-tty-on-linux-and-how-to-use-the-tty-command/
  For understanding the different options (and decide whether to select them, I will explain some tradeoffs below) see
  - CONFIG_UNIX98_PTYS - this is the  /dev/ptmx  / dev/ttyXX  /dev/pts/YY scheme.
  - CONFIG_LEGACY_PTYS - this is the legacy (BSD-like) scheme  using  /dev/ttyXX for the slaves (e.g. "physical terminals") and /dev/ptyYY for the master (e.g. what is usually pts). It is more conservativ
e in memory than the UNIX98_PTYS , but really no one uses them, so you can leave it (or use just it) for a programming exercise mostly.

 TTY n -> y
+LDISC_AUTOLOAD n
+LEGACY_PTYS n
+MWAVE n
+NULL_TTY n
+SERIAL_8250 n
+SERIAL_ALTERA_JTAGUART n
+SERIAL_ALTERA_UART n
+SERIAL_ARC n
+SERIAL_FSL_LINFLEXUART n
+SERIAL_FSL_LPUART n
+SERIAL_NONSTANDARD n
+SERIAL_SCCNXP n
+SERIAL_TIMBERDALE n
+SERIAL_UARTLITE n
+TRACE_SINK n
+TTY_PRINTK n
+UNIX98_PTYS n
+VT n

CONFIG_VT
Enables the virtual terminals (e.g. alt+ctrl+F3, setterm etc.), however the (surprisingly) important thing is that you MUST enable it if you want to use the keyobard and monitor.  Otherwise (if you don't need to switch VT's also) - you will need a Serial Console