Added INPUT_KEYBOARD and KEYBOARD_ATKBD . Now you can type, and console I/O is complete!
This concludes the "Monitor/Keyboard" setup. You can work with it in a QEMU window - as long as you provide it with a relevant init.
(Not done here, only in the debugging additions: you can actually test it with other mechanisms if you add to the kernel KDB and keyboard, MagicSysRQ key,...)

 INPUT_KEYBOARD n -> y
 SERIO n -> y
+KEYBOARD_ATKBD y
+KEYBOARD_LKKBD n
+KEYBOARD_NEWTON n
+KEYBOARD_OPENCORES n
+KEYBOARD_STOWAWAY n
+KEYBOARD_SUNKBD n
+KEYBOARD_XTKBD n
+SERIO_ALTERA_PS2 n
+SERIO_ARC_PS2 n
+SERIO_CT82C710 n
+SERIO_I8042 y
+SERIO_LIBPS2 y
+SERIO_PS2MULT n
+SERIO_RAW n
+SERIO_SERPORT n
+USERIO n