Adding FB_SIMPLE
The previous commit simply added interfaces, and support for the logo inclusion, but has no implementation of anything really.
This commit is not so different in terms of "results" - however, it implements features every  framebuffer device would implement (and most probably hardware accelerate) such as memory area copying, rectangle filling, and image blitting.

 FB_SIMPLE n -> y
+FB_CFB_COPYAREA y
+FB_CFB_FILLRECT y
+FB_CFB_IMAGEBLIT y